Add module creation feature with UI improvements
Implemented a new feature to create modules with a user-friendly UI in the VTL Data Generator. The new module includes a modal for user interactions, options to add multi-file uploaders, and a progress bar to track module creation progress. Additionally, updated styles, scripts, and configuration files to support the new functionality.

Bumped Version to 5.9